Frontend demo for fetching from database
